+++ changelog.md
@@ -39,6 +39,7 @@
 * #384 Web API: Inserting sluged recources like products require an URL record
 * #382 Promotion feed plugins: Asynchronous feed creation, more options and improvements
 * #433 GMC feed: Option to filter config records that have not been edited
+* #362 Display 'from {0}' for products with variant attributes
 
 ###Bugfixes###
 * Twitter Auth: fixed _SecurityTransparent_ error